requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84

pymupdf==1.24.9
pdfplumber==0.11.4

paddleocr==2.8.1
# Windows CPU PaddlePaddle wheels are hosted by Paddle:
paddlepaddle -f https://www.paddlepaddle.org.cn/whl/windows/mkl/avx/stable.html
